Rename the file that implements HKDF
