+++ src/lib.rs
@@ -48,8 +48,8 @@ mod aes;
 mod drbg;
 mod ecc;
 mod hash;
+mod hkdf;
 mod hmac;
-mod kdf;
 mod pbkdf2;
 mod rsa;
 mod sp800_108;
